[Refactor] Move query/fragment runtime state to ExecRuntime
