+++ be/src/exec/pipeline/pipeline_driver.cpp
@@ -37,14 +37,14 @@
 #include "exec/pipeline/primitives/driver_observer.h"
 #include "exec/pipeline/primitives/driver_queue.h"
 #include "exec/pipeline/primitives/event.h"
-#include "exec/pipeline/primitives/fragment_runtime_state.h"
 #include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/primitives/pipeline_observer.h"
-#include "exec/pipeline/primitives/query_runtime_state.h"
 #include "exec/pipeline/scan/morsel_queue.h"
 #include "exec/pipeline/scan/split_morsel_ticket_checker.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
 #include "exec/pipeline/source_operator.h"
+#include "exec/runtime/fragment_runtime_state.h"
+#include "exec/runtime/query_runtime_state.h"
 #include "gen_cpp/InternalService_types.h"
 #include "gutil/casts.h"
 #include "runtime/current_thread.h"
