+++ be/src/exec/pipeline/scan/olap_chunk_source.cpp
@@ -35,12 +35,12 @@
 #include "exec/olap_scan_node.h"
 #include "exec/olap_scan_prepare.h"
 #include "exec/pipeline/fragment_context.h"
-#include "exec/pipeline/primitives/query_runtime_state.h"
 #include "exec/pipeline/scan/glm_manager.h"
 #include "exec/pipeline/scan/olap_scan_context.h"
 #include "exec/pipeline/scan/scan_morsel.h"
 #include "exec/pipeline/scan/scan_operator.h"
 #include "exec/query_scan_metrics.h"
+#include "exec/runtime/query_runtime_state.h"
 #include "exprs/chunk_predicate_evaluator.h"
 #include "exprs/jsonpath.h"
 #include "gen_cpp/Metrics_types.h"
